# docs/api/gateway.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PHS — API pública (gateway)",
    "version": "0.1.0",
    "description": "Contrato que usa la aplicación web. Toda ruta entra por el gateway, que valida la sesión y reenvía a los servicios."
  },
  "paths": {
...
    "/api/v1/admin/users": {
      "get": {
        "summary": "Listar usuarios con sus roles",
        "tags": [
          "Administración"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "description": "Requiere un usuario administrador.",
        "responses": {
          "200": {
            "description": "Usuarios.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "email": {
                        "type": "string"
                      },
                      "displayName": {
                        "type": "string"
                      },
                      "active": {
                        "type": "boolean"
                      },
                      "isAdmin": {
                        "type": "boolean"
                      },
                      "memberships": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "practiceId": {
                              "type": "string",
                              "format": "uuid",
                              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                            },
                            "practiceName": {
                              "type": "string"
                            },
                            "role": {
                              "type": "string",
                              "enum": [
                                "pm",
                                "lead",
                                "director"
                              ],
                              "description": "Rol de un usuario en una práctica."
                            }
                          },
                          "required": [
                            "practiceId",
                            "practiceName",
                            "role"
                          ],
                          "additionalProperties": false
                        }
                      }
                    },
                    "required": [
                      "id",
                      "email",
                      "displayName",
                      "active",
                      "isAdmin",
                      "memberships"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión válida (`authentication_required`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso (`forbidden`), contraseña temporal sin cambiar (`password_change_required`) u origen no permitido (`origin_not_allowed`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "El servicio de identidad no respondió (`identity_unavailable`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear usuario con contraseña temporal",
        "tags": [
          "Administración"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "description": "Requiere un usuario administrador.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 320,
                    "pattern": "^[^\\s@]+@[^\\s@]+$"
                  },
                  "displayName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "isAdmin": {
                    "default": false,
                    "type": "boolean"
                  }
                },
                "required": [
                  "email",
                  "displayName"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "email": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "active": {
                          "type": "boolean"
                        },
                        "isAdmin": {
                          "type": "boolean"
                        },
                        "memberships": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "practiceId": {
                                "type": "string",
                                "format": "uuid",
                                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                              },
                              "practiceName": {
                                "type": "string"
                              },
                              "role": {
                                "type": "string",
                                "enum": [
                                  "pm",
                                  "lead",
                                  "director"
                                ],
                                "description": "Rol de un usuario en una práctica."
                              }
                            },
                            "required": [
                              "practiceId",
                              "practiceName",
                              "role"
                            ],
                            "additionalProperties": false
                          }
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "displayName",
                        "active",
                        "isAdmin",
                        "memberships"
                      ],
                      "additionalProperties": false
                    },
                    "temporaryPassword": {
                      "type": "string",
                      "description": "Se entrega una sola vez; no puede consultarse después."
                    }
                  },
                  "required": [
                    "user",
                    "temporaryPassword"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos (`invalid_request`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión válida (`authentication_required`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso (`forbidden`), contraseña temporal sin cambiar (`password_change_required`) u origen no permitido (`origin_not_allowed`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Conflicto (`email_taken`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "El servicio de identidad no respondió (`identity_unavailable`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{id}": {
      "patch": {
        "summary": "Cambiar nombre, estado o capacidad de administrador",
        "tags": [
          "Administración"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "description": "Requiere un usuario administrador.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "displayName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "active": {
                    "type": "boolean"
                  },
                  "isAdmin": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado. Al deshabilitarlo se revocan sus sesiones.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "email": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "active": {
                          "type": "boolean"
                        },
                        "isAdmin": {
                          "type": "boolean"
                        },
                        "memberships": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "practiceId": {
                                "type": "string",
                                "format": "uuid",
                                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                              },
                              "practiceName": {
                                "type": "string"
                              },
                              "role": {
                                "type": "string",
                                "enum": [
                                  "pm",
                                  "lead",
                                  "director"
                                ],
                                "description": "Rol de un usuario en una práctica."
                              }
                            },
                            "required": [
                              "practiceId",
                              "practiceName",
                              "role"
                            ],
                            "additionalProperties": false
                          }
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "displayName",
                        "active",
                        "isAdmin",
                        "memberships"
                      ],
                      "additionalProperties": false
                    },
                    "responsibilities": {
                      "type": "object",
                      "properties": {
                        "projectsAsPm": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "projectsAsLead": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "projectsAsTechnicalOwner": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "projectsAsPm",
                        "projectsAsLead",
                        "projectsAsTechnicalOwner"
                      ],
                      "additionalProperties": false,
                      "description": "Proyectos abiertos que necesitan nuevo responsable al deshabilitar al usuario."
                    }
                  },
                  "required": [
                    "user"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos (`invalid_request`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión válida (`authentication_required`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso (`forbidden`), contraseña temporal sin cambiar (`password_change_required`) u origen no permitido (`origin_not_allowed`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "No existe o está fuera del alcance del usuario (`not_found`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Conflicto (`last_admin`: debe quedar un administrador activo).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "El servicio de identidad no respondió (`identity_unavailable`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{id}/reset-password": {
      "post": {
        "summary": "Restablecer contraseña con un valor temporal",
        "tags": [
          "Administración"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "description": "Requiere un usuario administrador.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Contraseña temporal nueva; se revocan las sesiones del usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "temporaryPassword": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "temporaryPassword"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos (`invalid_request`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión válida (`authentication_required`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso (`forbidden`), contraseña temporal sin cambiar (`password_change_required`) u origen no permitido (`origin_not_allowed`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "No existe o está fuera del alcance del usuario (`not_found`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "El servicio de identidad no respondió (`identity_unavailable`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "phs_session",
        "description": "Cookie de sesión emitida por POST /api/v1/session."
      },
      "gatewaySignature": {
        "type": "apiKey",
        "in": "header",
        "name": "x-phs-internal-auth",
        "description": "Identidad firmada por el gateway (HMAC-SHA256, válida 60 s). No se acepta desde el navegador."
      }
    }
  }
}